+++ src/index.js
@@ -1,2 +1,1 @@
-module.exports = require("./core/proxy");
-module.exports.lookup = require("./core/lookup");
+module.exports = require("./plugin");